requirements.txt
numpy
pandas
matplotlib
python-dotenv
pymongo
certifi
pymongo[srv]==3.11
scikit-learn
pyaml
mlflow

# -e .
# while installing this pip install -r requirements.txt
# This will directly trigger setup.py to run
# Do this step at the last while deploying